requirements.txt
flask
pyotp
smartapi-python